requirements.txt
PySide6>=6.7.0
pypdf>=4.2.0
pydantic>=2.6.0
pdfminer.six>=20231228
requests>=2.31.0
